Custom widget areas for the website front page
Because we have a widget-driven custom content area on the front page, there may be some overlap with the widgets used on the sidebar and footers for the main part of the site. Four custom content areas are added to enable the site author to override the widgets used in these areas for the homepage of the site.

We also switch the homepage to only show the 3 most recent "news" posts