Hide header banner on single artwork posts
When the user clicks through to a specific work, that work shows up in the banner if it has been set as the featured image.
It is (a) weird to be showing it to them twice, (b) sub-optimal to be cramming it into the narrow space at the top of the page, and (c) very bizarre to show them a *different* image before the one they asked for (as happens when a banner is set but the featured image is not).

To keep the focus on the main artwork presentation, this prevents the header banner from rendering on single artwork pages.